Consolidate test files. Add test for negative constant double operator fix.
