+++ pyomo/repn/tests/gams/test_gams.py
@@ -13,10 +13,15 @@
 
 import os
 
-import pyutilib.th as unittest
+from six import StringIO
 
-from pyomo.environ import *
-import pyomo.opt
+import pyutilib.th as unittest
+from pyomo.core.base import NumericLabeler, SymbolMap
+from pyomo.environ import (Block, ConcreteModel, Connector, Constraint,
+                           Objective, TransformationFactory, Var, exp, log)
+from pyomo.repn.plugins.gams_writer import (StorageTreeChecker,
+                                            expression_to_string,
+                                            split_long_line)
 
 thisdir = os.path.dirname(os.path.abspath(__file__))
 
@@ -32,7 +37,7 @@ class Test(unittest.TestCase):
     def _get_fnames(self):
         class_name, test_name = self.id().split('.')[-2:]
         prefix = os.path.join(thisdir, test_name.replace("test_", "", 1))
-        return prefix+".gams.baseline", prefix+".gams.out"
+        return prefix + ".gams.baseline", prefix + ".gams.out"
 
     def _check_baseline(self, model, **kwds):
         baseline_fname, test_fname = self._get_fnames()
@@ -94,7 +99,7 @@ class Test(unittest.TestCase):
         components["con1"] = Constraint(expr=model.a >= 0)
         components["con2"] = Constraint(expr=model.a <= 1)
         components["con3"] = Constraint(expr=(0, model.a, 1))
-        components["con4"] = Constraint([1,2], rule=lambda m, i: model.a == i)
+        components["con4"] = Constraint([1, 2], rule=lambda m, i: model.a == i)
 
         for key in components:
             model.add_component(key, components[key])
@@ -107,7 +112,7 @@ class Test(unittest.TestCase):
         model.other = Block()
         model.other.a = Var()
         model.other.deactivate()
-        model.c = Constraint(expr=model.other.a + 2*model.x <= 0)
+        model.c = Constraint(expr=model.other.a + 2 * model.x <= 0)
         model.obj = Objective(expr=model.x)
         self._check_baseline(model)
 
@@ -123,28 +128,132 @@ class Test(unittest.TestCase):
 
         expr = M.abc**2.0
         self.assertEqual(str(expr), "abc**2.0")
-        self.assertEqual(expression_to_string(expr, tc, smap=smap), "power(abc, 2.0)")
+        self.assertEqual(expression_to_string(
+            expr, tc, smap=smap), "power(abc, 2.0)")
 
-        expr = log( M.abc**2.0 )
+        expr = log(M.abc**2.0)
         self.assertEqual(str(expr), "log(abc**2.0)")
-        self.assertEqual(expression_to_string(expr, tc, smap=smap), "log(power(abc, 2.0))")
+        self.assertEqual(expression_to_string(
+            expr, tc, smap=smap), "log(power(abc, 2.0))")
 
-        expr = log( M.abc**2.0 ) + 5
+        expr = log(M.abc**2.0) + 5
         self.assertEqual(str(expr), "log(abc**2.0) + 5")
-        self.assertEqual(expression_to_string(expr, tc, smap=smap), "log(power(abc, 2.0)) + 5")
+        self.assertEqual(expression_to_string(
+            expr, tc, smap=smap), "log(power(abc, 2.0)) + 5")
 
-        expr = exp( M.abc**2.0 ) + 5
+        expr = exp(M.abc**2.0) + 5
         self.assertEqual(str(expr), "exp(abc**2.0) + 5")
-        self.assertEqual(expression_to_string(expr, tc, smap=smap), "exp(power(abc, 2.0)) + 5")
+        self.assertEqual(expression_to_string(
+            expr, tc, smap=smap), "exp(power(abc, 2.0)) + 5")
 
-        expr = log( M.abc**2.0 )**4
+        expr = log(M.abc**2.0)**4
         self.assertEqual(str(expr), "log(abc**2.0)**4")
-        self.assertEqual(expression_to_string(expr, tc, smap=smap), "power(log(power(abc, 2.0)), 4)")
+        self.assertEqual(expression_to_string(
+            expr, tc, smap=smap), "power(log(power(abc, 2.0)), 4)")
 
-        expr = log( M.abc**2.0 )**4.5
+        expr = log(M.abc**2.0)**4.5
         self.assertEqual(str(expr), "log(abc**2.0)**4.5")
-        self.assertEqual(expression_to_string(expr, tc, smap=smap), "log(power(abc, 2.0)) ** 4.5")
+        self.assertEqual(expression_to_string(
+            expr, tc, smap=smap), "log(power(abc, 2.0)) ** 4.5")
 
+    def test_power_function_to_string(self):
+        m = ConcreteModel()
+        m.x = Var()
+        lbl = NumericLabeler('x')
+        smap = SymbolMap(lbl)
+        tc = StorageTreeChecker(m)
+        self.assertEqual(expression_to_string(
+            m.x ** -3, tc, lbl, smap=smap), "power(x1, (-3))")
+        self.assertEqual(expression_to_string(
+            m.x ** 0.33, tc, smap=smap), "x1 ** 0.33")
+        self.assertEqual(expression_to_string(
+            pow(m.x, 2), tc, smap=smap), "power(x1, 2)")
+
+    def test_fixed_var_to_string(self):
+        m = ConcreteModel()
+        m.x = Var()
+        m.y = Var()
+        m.z = Var()
+        m.z.fix(-3)
+        lbl = NumericLabeler('x')
+        smap = SymbolMap(lbl)
+        tc = StorageTreeChecker(m)
+        self.assertEqual(expression_to_string(
+            m.x + m.y - m.z, tc, lbl, smap=smap), "x1 + x2 + (-1)*(-3)")
+        m.z.fix(-400)
+        self.assertEqual(expression_to_string(
+            m.z + m.y - m.z, tc, smap=smap), "(-400) + x2 + (-1)*(-400)")
+        m.z.fix(8.8)
+        self.assertEqual(expression_to_string(
+            m.x + m.z - m.y, tc, smap=smap), "x1 + (8.8) + (-1)*x2")
+        m.z.fix(-8.8)
+        self.assertEqual(expression_to_string(
+            m.x * m.z - m.y, tc, smap=smap), "x1*(-8.8) + (-1)*x2")
+
+    def test_gams_connector_in_active_constraint(self):
+        m = ConcreteModel()
+        m.b1 = Block()
+        m.b2 = Block()
+        m.b1.x = Var()
+        m.b2.x = Var()
+        m.b1.c = Connector()
+        m.b1.c.add(m.b1.x)
+        m.b2.c = Connector()
+        m.b2.c.add(m.b2.x)
+        m.c = Constraint(expr=m.b1.c == m.b2.c)
+        m.o = Objective(expr=m.b1.x)
+        os = StringIO()
+        with self.assertRaises(RuntimeError):
+            m.write(os, format="gams")
+
+    def test_gams_expanded_connectors(self):
+        m = ConcreteModel()
+        m.x = Var()
+        m.y = Var()
+        m.CON1 = Connector()
+        m.CON1.add(m.x, 'v')
+        m.CON2 = Connector()
+        m.CON2.add(m.y, 'v')
+        m.c = Constraint(expr=m.CON1 + m.CON2 >= 10)
+        TransformationFactory("core.expand_connectors").apply_to(m)
+        m.o = Objective(expr=m.x)
+        os = StringIO()
+        io_options = dict(symbolic_solver_labels=True)
+        m.write(os, format="gams", io_options=io_options)
+        # no error if we're here, but check for some identifying string
+        self.assertIn("x + y", os.getvalue())
+
+    def test_split_long_line(self):
+        pat = "var1 + log(var2 / 9) - "
+        line = (pat * 10000) + "x"
+        self.assertEqual(split_long_line(line),
+                         pat * 3478 + "var1 +\nlog(var2 / 9) - " +
+                         pat * 3477 + "var1 +\nlog(var2 / 9) - " +
+                         pat * 3043 + "x")
+
+    def test_solver_arg(self):
+        m = ConcreteModel()
+        m.x = Var()
+        m.c = Constraint(expr=m.x == 2)
+        m.o = Objective(expr=m.x)
+        os = StringIO()
+        m.write(os, format="gams", io_options=dict(solver="gurobi"))
+        self.assertIn("option lp=gurobi", os.getvalue())
+
+    def test_negative_float_double_operator(self):
+        m = ConcreteModel()
+        m.x = Var()
+        m.y = Var()
+        m.z = Var(bounds=(0, 6))
+        m.c = Constraint(expr=(m.x * m.y * -2) == 0)
+        m.o = Objective(expr=m.z)
+        m.y.fix(-7)
+        m.x.fix(4)
+        lbl = NumericLabeler('x')
+        smap = SymbolMap(lbl)
+        tc = StorageTreeChecker(m)
+        self.assertEqual(expression_to_string(
+            m.x * m.z - m.y, tc, smap=smap), "(4)*x1 + (-1)*(-7)")
 
 
 class TestGams_writer(unittest.TestCase):
@@ -158,7 +267,7 @@ class TestGams_writer(unittest.TestCase):
     def _get_fnames(self):
         class_name, test_name = self.id().split('.')[-2:]
         prefix = os.path.join(thisdir, test_name.replace("test_", "", 1))
-        return prefix+".gams.baseline", prefix+".gams.out"
+        return prefix + ".gams.baseline", prefix + ".gams.out"
 
     def test_var_on_other_model(self):
         other = ConcreteModel()
@@ -166,7 +275,7 @@ class TestGams_writer(unittest.TestCase):
 
         model = ConcreteModel()
         model.x = Var()
-        model.c = Constraint(expr=other.a + 2*model.x <= 0)
+        model.c = Constraint(expr=other.a + 2 * model.x <= 0)
         model.obj = Objective(expr=model.x)
 
         baseline_fname, test_fname = self._get_fnames()
@@ -179,14 +288,14 @@ class TestGams_writer(unittest.TestCase):
     def test_var_on_nonblock(self):
         class Foo(Block().__class__):
             def __init__(self, *args, **kwds):
-                kwds.setdefault('ctype',Foo)
-                super(Foo,self).__init__(*args, **kwds)
+                kwds.setdefault('ctype', Foo)
+                super(Foo, self).__init__(*args, **kwds)
 
         model = ConcreteModel()
         model.x = Var()
         model.other = Foo()
         model.other.a = Var()
-        model.c = Constraint(expr=model.other.a + 2*model.x <= 0)
+        model.c = Constraint(expr=model.other.a + 2 * model.x <= 0)
         model.obj = Objective(expr=model.x)
 
         baseline_fname, test_fname = self._get_fnames()
@@ -195,6 +304,7 @@ class TestGams_writer(unittest.TestCase):
             RuntimeError,
             model.write, test_fname, format='gams')
         self._cleanup(test_fname)
+
 
 if __name__ == "__main__":
     unittest.main()
